+++ scripts/changelog.ts
@@ -138,7 +138,8 @@ async function exec(): Promise<void> {
         '**core**. Handle focus in nested documents correctly',
         '**plugin-rows**. Align top and bottom add butons correctly in custom integrations',
         '**plugin-text**. Disable slate undo/redo',
-        '**plugin-text**. Refocus after undo/redo'
+        '**plugin-text**. Refocus after undo/redo',
+        '**plugin-text**. Display Link-Overlay only when plugin is focused'
       ],
       internal: [
         '**demo**. Plugins available to the demo storybook are now defined in `demo/src/plugins.tsx`',